Lesson-1/Assignment 4/requirements.txt
fastapi
uvicorn
openai
pinecone
pydantic
python-dotenv
python-multipart
pypdf